requirements.txt
playwright
pytest-playwright
pytest-xdist
pytest-reporter-html1
